requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
h11==0.14.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
python-engineio==4.7.1
python-socketio==5.9.0
simple-websocket==0.10.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
wsproto==1.2.0
zipp==3.16.2
